requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14,<4.0
django-allauth>=0.51,<1.0
django-cors-headers>=3.10,<4.0
google-auth>=2.15,<3.0
